Module tremor_runtime::onramp::http

HTTP/REST Onramp

The http onramp process each POST request

Configuration

See Config for details.

Example

    - onramp::http:
        port: 8081
        host: "127.0.0.1"
        tls_cert: cert.pem
        tls_key: key.pem

Endpoints

PUT /raw

Takes the PUT body and puts it into the pipeline

PUT /v1

Takes the PUT body as a Tremor wrapped message and puts it on the pipeline.

Content Format

Tremor expects a PUT request with the Content-Type header set to application-json. Envelop encoding is, as the content type suggests a JSON.

The envelop has the structure:

{
  "header": {
    "uuid": "<base64 encoded raw uuid>",
    "parentUuid": "<base64 encoded raw uuid>" | null, // optional
    "contentType": "JSON" // other formats may be added.
  },
  "body": "<base64 encoded raw message>"
}

An example would be:

{
  "header": {
    "contentType":"JSON",
    "uuid":"gqB1SNkkEei6oAJCrBEAAg==",
    "parentUuid": null
  },
  "body":"eyJrZXkiOjQyfQ=="
}

Structs

Config

Onramp

Functions

perhaps_uuid_from_base64

uuid_from_base64

vec_u8_from_base64