Module tremor_runtime::op::grouping::bucket

Dimension based grouping with sliding windows

A This grouper is configred with a number of buckets and their alotted throughput on a time basis.

Messages are not combined by key and the alottment is applied in a sliding window fashion with a window granularity (a default of 10ms).

There is no 'magical' default bucket but one can be configured if desired along with a default rule.

Configuration

See Config for details.

Outputs

The 1st additional output is used to route data that was decided to be discarded to.

Example

- grouper::bucket:
    class1: # 1k messages per seconds on second granularity
      rate: 1000
    class2: # 500 messages per seconds on 10 second granularity
      rate: 5000
      time_range: 10000

Pseudocode

for e in events {
  if let bucket = buckets.get(e.classification) {
     dimension = bucket.keys.map(|key| e.data.get(key));
     if let window bucket.windows(dimension) {
       if window.inc() { pass } else { drop }
     } else {
       bucket.windows[dimension] = Window::new(bucket.limit)
       if bucket.windows[dimension].inc() { pass } else { drop }
     }
  } else {
    return drop
  }
}

Structs

Config

Grouper
Window based grouping

Rate
Single bucket specification