Module tremor_runtime::op::limiting::percentile

Percentile based sampling limiter

If adjustment is set this limiter will move the sampling percentage to keep the returned value from downstream operations between upper and lower limit.

Configuration

See Config for details.

Outputs

The 1st additional output is used to route data that was decided to be discarded.

Structs

Config
A Limiter algorithm that just lets trough a percentage of messages

Limiter